reworked parsing to do code seperately and moved markdown functions to seperate file
