+++ src/elements/mod.rs
@@ -6,6 +6,7 @@ mod element;
 mod layout;
 mod line_offset;
 mod loops;
+mod markdown;
 mod path;
 mod reuse;
 mod special;
