+++ src/elements/text.rs
@@ -2,6 +2,7 @@ use super::SvgElement;
 use crate::geometry::LocSpec;
 use crate::types::{attr_split_cycle, fstr, strp};
 
+use crate::elements::markdown::{get_md_value, TextClass};
 use crate::errors::{Result, SvgdxError};
 
 fn get_text_value(element: &mut SvgElement) -> String {
@@ -9,291 +10,6 @@ fn get_text_value(element: &mut SvgElement) -> String {
         .pop_attr("text")
         .expect("no text attr in process_text_attr");
     text_string(&text_value)
-}
-
-fn get_md_value(element: &mut SvgElement) -> (Vec<String>, Vec<u32>) {
-    let text_value = element
-        .pop_attr("md")
-        .expect("no md attr in process_text_attr");
-
-    let (parsed_string, sections) = md_parse(&text_value);
-
-    let mut state_per_char = vec![0; parsed_string.len()];
-
-    for s in sections {
-        let bit = s.code_bold_italic;
-        for i in state_per_char.iter_mut().take(s.end_ind).skip(s.start_ind) {
-            *i |= 1 << bit;
-        }
-    }
-
-    let mut strings = vec![];
-    let mut states = vec![];
-    for i in 0..parsed_string.len() {
-        if i == 0 || states[states.len() - 1] != state_per_char[i] {
-            strings.push(String::new());
-            states.push(state_per_char[i])
-        }
-        strings
-            .last_mut()
-            .expect("filled from i == 0")
-            .push(parsed_string[i]);
-    }
-
-    (strings, states)
-}
-
-#[derive(Debug)]
-struct SectionData {
-    start_ind: usize,
-    end_ind: usize,
-    code_bold_italic: u32,
-}
-
-// based on the commonmarkdown implementation
-#[derive(Debug)]
-struct DelimiterData {
-    ind: usize, // goes just before this char
-    char_type: char,
-    num_delimiters: usize,
-    is_active: bool,
-    could_open: bool,
-    could_close: bool,
-}
-
-fn md_parse_escapes_and_delimiters(text_value: &str) -> (Vec<char>, Vec<DelimiterData>) {
-    let mut result = vec![];
-    let mut delimiters = vec![DelimiterData {
-        ind: 0,
-        char_type: ' ',
-        num_delimiters: 0,
-        is_active: false,
-        could_open: false,
-        could_close: false,
-    }];
-    let mut escaped = false;
-
-    // first pass process \ and find delimiters
-    for c in text_value.chars() {
-        let mut add = true;
-        match (c, escaped) {
-            ('\\', false) => {
-                add = false;
-                escaped = true;
-            }
-            ('\\', true) => {
-                escaped = false;
-            }
-            // the delimiters
-            ('`', false) | ('_', false) | ('*', false) => {
-                let last = delimiters.last_mut().expect("garenteed not to be empty");
-                if c == last.char_type && last.ind == result.len() {
-                    // is a continuation
-                    last.num_delimiters += 1;
-                } else {
-                    delimiters.push(DelimiterData {
-                        ind: result.len(),
-                        char_type: c,
-                        num_delimiters: 1,
-                        is_active: true,
-                        could_open: true,
-                        could_close: true,
-                    });
-                }
-                add = false;
-            }
-            ('`', true) | ('_', true) | ('*', true) => {
-                escaped = false;
-            }
-            ('n', true) => {
-                add = false;
-                result.push('\n');
-                escaped = false;
-            }
-            (_, true) => {
-                // was not an escape
-                result.push('\\');
-                escaped = false;
-            }
-            (_, false) => {}
-        }
-        if add {
-            result.push(c);
-        }
-    }
-
-    (result, delimiters)
-}
-
-fn md_parse_set_delimiter_open_close(result: &[char], delimiters: &mut [DelimiterData]) {
-    // set could open/close
-    for i in 0..delimiters.len() {
-        let prev_char;
-        let next_char;
-        if i != 0 && delimiters[i - 1].ind == delimiters[i].ind {
-            prev_char = delimiters[i - 1].char_type;
-        } else if delimiters[i].ind == 0 {
-            prev_char = ' ';
-        } else {
-            prev_char = result[delimiters[i].ind - 1];
-        }
-
-        if i != delimiters.len() - 1 && delimiters[i + 1].ind == delimiters[i].ind {
-            next_char = delimiters[i + 1].char_type;
-        } else if delimiters[i].ind == result.len() {
-            next_char = ' ';
-        } else {
-            next_char = result[delimiters[i].ind];
-        }
-
-        match (prev_char.is_whitespace(), next_char.is_whitespace()) {
-            (false, false) => {
-                if delimiters[i].char_type == '_' {
-                    delimiters[i].could_open = false;
-                    delimiters[i].could_close = false;
-                }
-            }
-            (true, false) => {
-                delimiters[i].could_close = false;
-            }
-            (false, true) => {
-                delimiters[i].could_open = false;
-            }
-            (true, true) => {
-                delimiters[i].could_open = false;
-                delimiters[i].could_close = false;
-            }
-        }
-
-        if next_char.is_ascii_punctuation()
-            && (!prev_char.is_whitespace() || !prev_char.is_ascii_punctuation())
-        {
-            delimiters[i].could_open = false;
-        }
-        if prev_char.is_ascii_punctuation()
-            && (!next_char.is_whitespace() || !next_char.is_ascii_punctuation())
-        {
-            delimiters[i].could_close = false;
-        }
-    }
-}
-
-fn md_parse_eval_sections(delimiters: &mut [DelimiterData]) -> Vec<SectionData> {
-    let mut sections = vec![];
-    let stack_bottom = 0; // because I have a null element in it
-    let mut current_position = stack_bottom + 1;
-    let mut opener_a = [stack_bottom; 3];
-    let mut opener_d = [stack_bottom; 3];
-    let mut opener_t = [stack_bottom; 3];
-
-    loop {
-        while current_position != delimiters.len()
-            && !delimiters[current_position].could_close
-            && delimiters[current_position].is_active
-        {
-            current_position += 1;
-        }
-        if current_position == delimiters.len() {
-            break;
-        }
-        let opener_min = match delimiters[current_position].char_type {
-            '*' => &mut opener_a,
-            '_' => &mut opener_d,
-            '`' => &mut opener_t,
-            _ => panic!("this cant happen as current_position starts at 0 and all other delimiters are of above types"),
-        };
-
-        let min = opener_min[delimiters[current_position].num_delimiters % 3].max(stack_bottom);
-        let mut opener_ind = current_position - 1;
-        while opener_ind > min {
-            // found opener
-            if delimiters[opener_ind].is_active
-                && delimiters[opener_ind].could_open
-                && delimiters[opener_ind].char_type == delimiters[current_position].char_type
-            {
-                if (delimiters[opener_ind].could_close || delimiters[current_position].could_open)
-                    && delimiters[opener_ind].num_delimiters % 3
-                        != delimiters[current_position].num_delimiters % 3
-                {
-                } else {
-                    // found valid opener
-                    break;
-                }
-            }
-            opener_ind -= 1;
-        }
-
-        if opener_ind == min {
-            // not found a opener
-            opener_min[delimiters[current_position].num_delimiters % 3] = current_position - 1;
-            current_position += 1;
-        } else {
-            delimiters[current_position].could_open = false;
-            delimiters[opener_ind].could_close = false;
-            // did
-            let code = delimiters[current_position].char_type == '`';
-            let strong = !code
-                && delimiters[opener_ind].num_delimiters >= 2
-                && delimiters[current_position].num_delimiters >= 2;
-            sections.push(SectionData {
-                start_ind: delimiters[opener_ind].ind,
-                end_ind: delimiters[current_position].ind,
-                code_bold_italic: match (code, strong) {
-                    (true, _) => 0,
-                    (_, true) => 1,
-                    (_, _) => 2,
-                },
-            });
-
-            delimiters[opener_ind].num_delimiters -= 1 + (strong as usize);
-            delimiters[current_position].num_delimiters -= 1 + (strong as usize);
-
-            if delimiters[opener_ind].num_delimiters == 0 {
-                delimiters[opener_ind].is_active = false;
-            }
-            if delimiters[current_position].num_delimiters == 0 {
-                delimiters[current_position].is_active = false;
-                current_position += 1;
-            }
-
-            for d in &mut delimiters[(opener_ind + 1)..current_position] {
-                d.is_active = false;
-            }
-        }
-    }
-    sections
-}
-
-fn md_parse(text_value: &str) -> (Vec<char>, Vec<SectionData>) {
-    let (mut result, mut delimiters) = md_parse_escapes_and_delimiters(text_value);
-    md_parse_set_delimiter_open_close(&result, &mut delimiters);
-    let mut sections = md_parse_eval_sections(&mut delimiters);
-
-    let mut final_result = vec![];
-
-    // work from the back to avoid index invalidation
-    for d in delimiters.into_iter().rev() {
-        while d.ind < result.len() {
-            if let Some(thing) = result.pop() {
-                final_result.push(thing);
-            }
-        }
-
-        for s in sections.iter_mut() {
-            // if start needs to be after or equal
-            if s.start_ind >= d.ind {
-                s.start_ind += d.num_delimiters;
-            }
-            if s.end_ind > d.ind {
-                // if end needs to be after
-                s.end_ind += d.num_delimiters;
-            }
-        }
-        let mut temp = vec![d.char_type; d.num_delimiters];
-        final_result.append(&mut temp);
-    }
-
-    (final_result.into_iter().rev().collect(), sections)
 }
 
 /// Convert unescaped r"\n" into newline characters for multi-line text
@@ -454,8 +170,17 @@ pub fn process_text_attr(element: &SvgElement) -> Result<(SvgElement, Vec<SvgEle
             "has both attributes of text and md".to_owned(),
         ));
     } else if orig_elem.get_attr("text").is_some() {
-        text_values = vec![get_text_value(&mut orig_elem)];
-        state_values = vec![0];
+        if orig_elem.has_class("d-markdown") {
+            // as to call must have one of them
+            (text_values, state_values) = get_md_value(&mut orig_elem);
+        } else {
+            text_values = vec![get_text_value(&mut orig_elem)];
+            state_values = vec![TextClass {
+                code: false,
+                bold: false,
+                italic: false,
+            }];
+        }
     } else {
         // as to call must have one of them
         (text_values, state_values) = get_md_value(&mut orig_elem);
@@ -507,7 +232,11 @@ pub fn process_text_attr(element: &SvgElement) -> Result<(SvgElement, Vec<SvgEle
     for i in 0..lines.len() {
         if lines[i].is_empty() {
             lines[i].push("");
-            line_types[i].push(0);
+            line_types[i].push(TextClass {
+                code: false,
+                bold: false,
+                italic: false,
+            });
         }
     }
     let line_count = lines.len();
@@ -586,13 +315,13 @@ pub fn process_text_attr(element: &SvgElement) -> Result<(SvgElement, Vec<SvgEle
     }
 
     if !multielement {
-        if line_types[0][0] & (1 << 0) != 0 {
+        if line_types[0][0].code {
             text_classes.push("d-text-monospace".to_string());
         }
-        if line_types[0][0] & (1 << 1) != 0 {
+        if line_types[0][0].bold {
             text_classes.push("d-text-bold".to_string());
         }
-        if line_types[0][0] & (1 << 2) != 0 {
+        if line_types[0][0].italic {
             text_classes.push("d-text-italic".to_string());
         }
     }
@@ -668,13 +397,13 @@ pub fn process_text_attr(element: &SvgElement) -> Result<(SvgElement, Vec<SvgEle
                     }
                 }
 
-                if line_types[idx][idn] & (1 << 0) != 0 {
+                if line_types[idx][idn].code {
                     tspan.add_class("d-text-monospace");
                 }
-                if line_types[idx][idn] & (1 << 1) != 0 {
+                if line_types[idx][idn].bold {
                     tspan.add_class("d-text-bold");
                 }
-                if line_types[idx][idn] & (1 << 2) != 0 {
+                if line_types[idx][idn].italic {
                     tspan.add_class("d-text-italic");
                 }
 
@@ -733,122 +462,5 @@ mod tests {
         // when precedes '\n', '\' escapes it.
         let text = r"Hello, \\nworld!";
         assert_eq!(text_string(text), r"Hello, \nworld!");
-    }
-
-    #[test]
-    fn test_md_parse() {
-        // the basic examples no actual md
-
-        let text = r"Hello, \nworld!";
-        assert_eq!(
-            format!("{:?}", md_parse(text)),
-            "(['H', 'e', 'l', 'l', 'o', ',', ' ', '\\n', 'w', 'o', 'r', 'l', 'd', '!'], [])"
-        );
-
-        // when not part of a '\n', '\' is not special
-        let text = r"Hello, world! \1";
-        assert_eq!(format!("{:?}",md_parse(text)), "(['H', 'e', 'l', 'l', 'o', ',', ' ', 'w', 'o', 'r', 'l', 'd', '!', ' ', '\\\\', '1'], [])");
-
-        // when precedes '\n', '\' escapes it.
-        let text = r"Hello, \\nworld!";
-        assert_eq!(
-            format!("{:?}", md_parse(text)),
-            "(['H', 'e', 'l', 'l', 'o', ',', ' ', '\\\\', 'n', 'w', 'o', 'r', 'l', 'd', '!'], [])"
-        );
-
-        fn sd(s: i32, e: i32, i: i32) -> String {
-            "SectionData { start_ind: ".to_owned()
-                + &s.to_string()
-                + ", end_ind: "
-                + &e.to_string()
-                + ", code_bold_italic: "
-                + &i.to_string()
-                + " }"
-        }
-
-        // using the md
-        let text = r"He*ll*o, \nworld!";
-        assert_eq!(
-            format!("{:?}", md_parse(text)),
-            "(['H', 'e', 'l', 'l', 'o', ',', ' ', '\\n', 'w', 'o', 'r', 'l', 'd', '!'], ["
-                .to_owned()
-                + &sd(2, 4, 2)
-                + "])"
-        );
-
-        // mismatched
-        let text = r"*Hello** , \nworld!";
-        assert_eq!(
-            format!("{:?}", md_parse(text)),
-            "(['H', 'e', 'l', 'l', 'o', '*', ' ', ',', ' ', '\\n', 'w', 'o', 'r', 'l', 'd', '!'], ["
-                .to_owned()
-                + &sd(0, 5, 2) + "])"
-        );
-
-        // diff type
-        let text = r"He*llo_, \nworld!";
-        assert_eq!(format!("{:?}",md_parse(text)), "(['H', 'e', '*', 'l', 'l', 'o', '_', ',', ' ', '\\n', 'w', 'o', 'r', 'l', 'd', '!'], [])");
-
-        // multiple diff type
-        let text = r"_hello*";
-        assert_eq!(
-            format!("{:?}", md_parse(text)),
-            "(['_', 'h', 'e', 'l', 'l', 'o', '*'], [])"
-        );
-
-        // multiple same type
-        let text = r"He*ll*o, \nw*or*ld!";
-        assert_eq!(
-            format!("{:?}", md_parse(text)),
-            "(['H', 'e', 'l', 'l', 'o', ',', ' ', '\\n', 'w', 'o', 'r', 'l', 'd', '!'], ["
-                .to_owned()
-                + &sd(2, 4, 2)
-                + ", "
-                + &sd(9, 11, 2)
-                + "])"
-        );
-
-        // space before
-        let text = r"**foo bar **";
-        assert_eq!(
-            format!("{:?}", md_parse(text)),
-            "(['*', '*', 'f', 'o', 'o', ' ', 'b', 'a', 'r', ' ', '*', '*'], [])"
-        );
-
-        // punctuation before alphnum after
-        let text = r"**(**foo)";
-        assert_eq!(
-            format!("{:?}", md_parse(text)),
-            "(['*', '*', '(', '*', '*', 'f', 'o', 'o', ')'], [])"
-        );
-    }
-
-    #[test]
-    fn test_get_md_value() {
-        let mut el = SvgElement::new("text", &[]);
-        let text = r"foo";
-        el.set_attr("md", text);
-        assert_eq!(format!("{:?}", get_md_value(&mut el)), "([\"foo\"], [0])");
-
-        let text = r"**(**foo)";
-        el.set_attr("md", text);
-        assert_eq!(
-            format!("{:?}", get_md_value(&mut el)),
-            "([\"**(**foo)\"], [0])"
-        );
-
-        let text = r"*foo *bar**";
-        el.set_attr("md", text);
-        assert_eq!(
-            format!("{:?}", get_md_value(&mut el)),
-            "([\"foo bar\"], [4])"
-        );
-
-        let text = r"*foo**bar**baz*";
-        el.set_attr("md", text);
-        assert_eq!(
-            format!("{:?}", get_md_value(&mut el)),
-            "([\"foo\", \"bar\", \"baz\"], [4, 6, 4])"
-        );
     }
 }
